requirements.txt
langchain>=0.3
langchain-core>=0.3
langchain-community>=0.3
langgraph
chromadb
openai>=1.50
tiktoken
google-search-results
duckduckgo-search
beautifulsoup4
readability-lxml
requests
python-dotenv
streamlit>=1.28.0
